django-learning/myshop/pyproject.toml
[project]
name = "myshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "sentry-sdk[django] (>=2.47.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "markdown (>=3.10,<4.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
coverage = "^7.11.3"

